requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
dj-rest-auth
django-allauth[with_social]
djangorestframework-simplejwt
drf-yasg
drf-nested-routers
django-friendship